Add reset

+++ wasm/web/main.js
@@ -9,30 +9,42 @@ const print = (value) => {
 	}
 }
 
+const init = () => {
+	const channel = new MessageChannel()
+	sw.postMessage({ type: 'init' }, [channel.port1])
+	worker.postMessage({ type: 'init' }, [channel.port2])
+}
+
+let sw
+let worker
+let controller
 const startServer = async () =>
 	new Promise((resolve) => {
 		print('Loading PalaCMS...')
-		const worker = new Worker('/worker.js')
-
-		const init = (sw) => {
-			const channel = new MessageChannel()
-			worker.postMessage({ type: 'init' }, [channel.port2])
-			sw.postMessage({ type: 'init' }, [channel.port1])
-			window.addEventListener('beforeunload', () => {
-				sw.postMessage({ type: 'close' })
-				worker.terminate()
-			})
-			fetch('/api/palacms/info').then(() => resolve())
-		}
-
+		if (worker) worker.terminate()
+		worker = new Worker('/worker.js')
 		worker.addEventListener('message', (event) => {
 			const { type, value } = event.data
 			if (type === 'ready') {
+				const channel = new MessageChannel()
+				worker.postMessage({ type: 'init' }, [channel.port2])
 				navigator.serviceWorker.register('/sw.js', { scope: '/' }).then((registration) => {
-					if (registration.active) {
-						init(registration.active)
-					}
-					registration.addEventListener('updatefound', () => init(registration.installing))
+					sw = registration.installing ?? registration.active
+					init()
+					registration.addEventListener('updatefound', () => {
+						sw = registration.installing
+						init()
+					})
+
+					controller = new AbortController()
+					resolve(
+						fetch('/api/palacms/info', { signal: controller.signal }).then((res) => {
+							controller = undefined
+							if (!res.ok) {
+								throw new Error('Non-ok response')
+							}
+						})
+					)
 				})
 			} else if (type === 'output') {
 				print(value)
@@ -40,28 +52,64 @@ const startServer = async () =>
 		})
 	})
 
-if (window.name === 'server') {
-	startServer()
-	document.querySelector('#info').open = false
-	document.querySelector('#open').hidden = false
-} else {
-	document.querySelector('#start').hidden = false
+const stopServer = () => {
+	if (sw) sw.postMessage({ type: 'close' })
+	if (worker) worker.terminate()
+	if (controller) controller.abort()
+	sw = undefined
+	worker = undefined
+	controller = undefined
 }
 
-let opened = false
-document.querySelector('#open').addEventListener('click', () => {
-	opened = true
-})
+const resetServer = async () => {
+	stopServer()
+	const root = await navigator.storage.getDirectory()
+	await root.removeEntry('pb_data', { recursive: true })
+}
+
+if (window.name === 'server') {
+	document.querySelector('#reset').hidden = false
+	startServer().then(() => {
+		document.querySelector('#open').hidden = false
+	})
+} else {
+	document.querySelector('#start').hidden = false
+	document.querySelector('#info').open = true
+	navigator.storage.getDirectory().then(async (root) => {
+		const keys = await Array.fromAsync(root.keys())
+		if (keys.includes('pb_data')) {
+			document.querySelector('#reset').hidden = false
+		}
+	})
+}
 
 document.querySelector('#start').addEventListener('click', () => {
 	window.name = 'server'
-	startServer().then(() => {
-		if (!opened) {
-			window.open(location.href, '_blank')
-		}
-		history.replaceState(null, '', '/')
-	})
 	document.querySelector('#info').open = false
-	document.querySelector('#open').hidden = false
 	document.querySelector('#start').hidden = true
+	document.querySelector('#reset').hidden = false
+	startServer()
+		.then(() => {
+			document.querySelector('#open').hidden = false
+			history.replaceState(null, '', '/')
+		})
+		.catch((error) => {
+			console.error(error)
+			print(error)
+			stopServer()
+		})
+})
+
+document.querySelector('#reset').addEventListener('click', () => {
+	resetServer().then(() => {
+		window.name = ''
+		output.innerHTML = ''
+		document.querySelector('#open').hidden = true
+		document.querySelector('#start').hidden = false
+		document.querySelector('#reset').hidden = true
+	})
+})
+
+window.addEventListener('beforeunload', () => {
+	stopServer()
 })
